pyproject.toml
[project]
name = "hitalent-test"
version = "0.1.0"
description = ""
authors = []
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi",
    "pydantic",
    "pydantic-settings",
    "uvicorn",
    "sqlalchemy",
    "asyncpg",
    "alembic",
    "python-dotenv",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"


[tool.poetry]
package-mode = false
